requirements.txt
beautifulsoup4>=4.12
lxml>=4.9
playwright>=1.44
tldextract>=3.6
orjson>=3.9; platform_python_implementation != 'PyPy'
tqdm>=4.66
transformers>=4.42
torch>=2.1
pyyaml>=6.0
numpy>=1.25
requests>=2.31
matplotlib>=3.8
seaborn>=0.13
scikit-learn>=1.3
lime>=0.2.0.1
shap>=0.43.0
